feat: integrated with main api; reverted LoadConfigFile refactor
