+++ api/push/push.go
@@ -1,15 +1,13 @@
-package main
+package push
 
 import (
-	"fmt"
 	"net/http"
 
 	"github.com/gin-gonic/gin"
-	"github.com/merico-dev/lake/push-api/db"
 )
 
 /*
-	POST /api/:tableName
+	POST /push/:tableName
 	[
 		{
 			"id": 1,
@@ -17,17 +15,6 @@ import (
 		}
 	]
 */
-
-func main() {
-	db.InitDb()
-	router := gin.Default()
-	router.POST("/api/:tableName", Post)
-
-	err := router.Run("localhost:9123")
-	if err != nil {
-		fmt.Println("ERROR >>> a2kj4: ", err)
-	}
-}
 
 func Post(c *gin.Context) {
 	var err error
@@ -39,7 +26,7 @@ func Post(c *gin.Context) {
 		_ = c.AbortWithError(http.StatusBadRequest, err)
 	}
 	for _, unknownThing := range unknownThings {
-		rowsAffected, err := db.InsertThing(tableName, unknownThing)
+		rowsAffected, err := InsertThing(tableName, unknownThing)
 		if err != nil {
 			_ = c.AbortWithError(http.StatusInternalServerError, err)
 		}
