+++ api/router.go
@@ -7,6 +7,7 @@ import (
 	"github.com/gin-gonic/gin"
 	"github.com/merico-dev/lake/api/env"
 	"github.com/merico-dev/lake/api/pipelines"
+	"github.com/merico-dev/lake/api/push"
 	"github.com/merico-dev/lake/api/task"
 	"github.com/merico-dev/lake/plugins/core"
 	"github.com/merico-dev/lake/services"
@@ -20,6 +21,7 @@ func RegisterRouter(r *gin.Engine) {
 	r.GET("/pipelines/:pipelineId/tasks", task.Index)
 	r.POST("/env", env.Set)
 	r.GET("/env", env.Get)
+	r.POST("/push/:tableName", push.Post)
 
 	// mount all api resources for all plugins
 	pluginsApiResources, err := services.GetPluginsApiResources()
