+++ config/config.go
@@ -33,19 +33,16 @@ type Config struct {
 
 var V *viper.Viper
 
-func LoadConfigFile(envPath string) *viper.Viper {
-	if envPath == "" {
-		envPath = ".env"
-	}
+func LoadConfigFile() *viper.Viper {
 	V = viper.New()
-	V.SetConfigFile(envPath)
+	V.SetConfigFile(".env")
 	_ = V.ReadInConfig()
 	V.AutomaticEnv()
 	return V
 }
 
 func init() {
-	V := LoadConfigFile("")
+	V := LoadConfigFile()
 	V.SetDefault("PORT", ":8080")
 	V.SetDefault("PLUGIN_DIR", "bin/plugins")
 	// This line is essential for reading and writing
