+++ plugins/gitlab/api/gitlab_sources.go
@@ -37,7 +37,7 @@ func PutSource(input *core.ApiResourceInput) (*core.ApiResourceOutput, error) {
 	if err != nil {
 		return nil, err
 	}
-	V := config.LoadConfigFile("")
+	V := config.LoadConfigFile()
 	if gitlabSource.GITLAB_ENDPOINT != "" {
 		V.Set("GITLAB_ENDPOINT", gitlabSource.GITLAB_ENDPOINT)
 	}
@@ -81,7 +81,7 @@ func GetSource(input *core.ApiResourceInput) (*core.ApiResourceOutput, error) {
 }
 
 func GetSourceFromEnv() (*GitlabResponse, error) {
-	V := config.LoadConfigFile("")
+	V := config.LoadConfigFile()
 	var configJson GitlabConfig
 	err := V.Unmarshal(&configJson)
 	if err != nil {
