+++ plugins/jenkins/api/jenkins_sources.go
@@ -51,7 +51,7 @@ func PutSource(input *core.ApiResourceInput) (*core.ApiResourceOutput, error) {
 	if err != nil {
 		return nil, err
 	}
-	V := config.LoadConfigFile("")
+	V := config.LoadConfigFile()
 	V.Set("JENKINS_ENDPOINT", jenkinsSource.JENKINS_ENDPOINT)
 	V.Set("JENKINS_USERNAME", jenkinsSource.JENKINS_USERNAME)
 	V.Set("JENKINS_PASSWORD", jenkinsSource.JENKINS_PASSWORD)
