--- push-api/Makefile
@@ -1,7 +1,0 @@
-build: 
-	go build -o ./bin/push-api
-
-run: 
-	./bin/push-api
-
-build-and-run: build run
